+++ lib/apave/apave.tcl
@@ -7,7 +7,7 @@
 ###########################################################
 
 package require Tk
-package provide apave 4.5.4
+package provide apave 4.5.5
 
 source [file join [file dirname [info script]] apavedialog.tcl]